Refactor extensions tests
